Moved style files location
